victory / death animations
Screen transitions are in, and player's can't attack during death time
(the player who is alive can still move around).

Animations for win/lose are in, but need some work. The spritesheets for
fairy godmother do not cut properly (they only get the wings for some
reason), and the wolf animations need tuning, not sure what it's meant
to look like. Also there is a weird 1-frame flash to the idle sprite
between loops ...